+++ Dx2dGameMKU/GameEngineContents/PlayerState_SpecialAttack_AxeKick.h
@@ -24,7 +24,7 @@ private:
 	static const std::string_view AniFileName;
 	static const std::pair<int, int> AniCutFrame;
 	static const float AniInterTime;
-	static const float Damage;
+	static const int Damage;
 
 	void LoadAnimation();
 	void CreateAnimation();
